Remove a bunch of dependencies.
Most of this is removing them from various ufront libs we have as dependencies.
eg. Random and CleverSort are no longer needed.

Removing Detox as it is only used in one place (js.ufront.web.context.HttpResponse)
and was easy to achieve the same result with native methods.

Leaving pushstate as a dependency for now, the functionality it provides is slightly non-trivial.